Included blocklist for JWTs and logic for logout.

+++ resources/item.py
@@ -1,6 +1,6 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 
 from model import ItemModel
 
@@ -20,6 +20,9 @@ class Item(MethodView):
     
     @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin priviledge required to delete item")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
